requirements.txt
# ============================================================
# Heri-Sci Backend Requirements
# Python 3.8+ required
# ============================================================

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Environment & Configuration
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# AI & Machine Learning
google-generativeai==0.3.2
openai==1.3.0

# HTTP & API
requests==2.31.0
urllib3==2.1.0

# Data Processing
pandas==2.1.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Optional: Firebase Admin (if using server-side Firebase)
# firebase-admin==6.2.0

# Optional: Advanced Image Processing
# scikit-image==0.22.0
# scipy==1.11.4

# Development Dependencies (optional)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0

# ============================================================
# Installation Instructions:
# ============================================================
# 1. Create virtual environment:
#    python -m venv venv
#
# 2. Activate virtual environment:
#    Windows: venv\Scripts\activate
#    macOS/Linux: source venv/bin/activate
#
# 3. Install dependencies:
#    pip install -r requirements.txt
#
# 4. Verify installation:
#    pip list
# ============================================================
